refactor: split misc_queries.rs into invite, ban, prekey, settings, sync, and permission modules
