+++ server-rs/src/db/mod.rs
@@ -7,7 +7,12 @@ mod channel_queries;
 mod member_queries;
 mod message_queries;
 mod role_queries;
-mod misc_queries;
+mod sync_queries;
+mod permission_queries;
+mod invite_queries;
+mod ban_queries;
+mod prekey_queries;
+mod settings_queries;
 mod dm_queries;
 mod thread_queries;
 mod reaction_queries;
@@ -25,7 +30,12 @@ pub use channel_queries::*;
 pub use member_queries::*;
 pub use message_queries::*;
 pub use role_queries::*;
-pub use misc_queries::*;
+pub use sync_queries::*;
+pub use permission_queries::*;
+pub use invite_queries::*;
+pub use ban_queries::*;
+pub use prekey_queries::*;
+pub use settings_queries::*;
 pub use dm_queries::*;
 pub use thread_queries::*;
 pub use reaction_queries::*;
